refactor(pages): migrate translations location to adjacent
