+++ src/routes/[[lang=lang]]/_page/translation/en.ts
