+++ src/routes/[[lang=lang]]/_page/translation/index.ts
