+++ src/routes/[[lang=lang]]/_page/translation/vi.ts
