+++ src/routes/[[lang=lang]]/code-of-conduct/_page/translation/en.ts
