+++ src/routes/[[lang=lang]]/events/+page.server.ts
@@ -2,9 +2,9 @@ import { mail } from '$shared/actions/mail/mail.server';
 import { translations as mailT } from '$shared/actions/mail/translation';
 import { createMockedEvents } from '$shared/mocks';
 import type { Language } from '$shared/services/i18n';
-import { translations as pageT } from '$shared/services/i18n/translations/pages/events';
 
 import type { PageServerLoadEvent, Actions } from './$types';
+import { translations as pageT } from './_page/translation';
 
 export async function load({ params }: PageServerLoadEvent) {
   const lang = params.lang as Language;
