+++ src/routes/[[lang=lang]]/events/[slug]/_page/translation/index.ts
