+++ src/routes/[[lang=lang]]/events/_page/translation/en.ts
