+++ src/routes/[[lang=lang]]/events/_page/translation/index.ts
