+++ src/routes/[[lang=lang]]/events/_page/translation/vi.ts
