+++ src/routes/[[lang=lang]]/impact/+page.server.ts
@@ -1,8 +1,8 @@
 import { createMockedProjects } from '$shared/mocks';
 import type { Language } from '$shared/services/i18n';
-import { translations } from '$shared/services/i18n/translations/pages/impact';
 
 import type { PageServerLoadEvent } from './$types';
+import { translations } from './_page/translation';
 
 export async function load({ params }: PageServerLoadEvent) {
   const lang = params.lang as Language;
