+++ src/routes/[[lang=lang]]/impact/_page/translation/en.ts
