+++ src/routes/[[lang=lang]]/impact/_page/translation/index.ts
