+++ src/routes/[[lang=lang]]/jobs/+page.server.ts
@@ -2,9 +2,9 @@ import { mail } from '$shared/actions/mail/mail.server';
 import { translations as tMail } from '$shared/actions/mail/translation';
 import { createMockedJobs } from '$shared/mocks';
 import type { Language } from '$shared/services/i18n';
-import { translations as tPage } from '$shared/services/i18n/translations/pages/jobs';
 
 import type { PageServerLoadEvent, Actions } from './$types';
+import { translations as tPage } from './_page/translation';
 
 export async function load({ params }: PageServerLoadEvent) {
   const lang = params.lang as Language;
