+++ src/routes/[[lang=lang]]/jobs/_page/translation/en.ts
