+++ src/routes/[[lang=lang]]/jobs/_page/translation/index.ts
