+++ src/routes/[[lang=lang]]/jobs/_page/translation/vi.ts
