+++ src/routes/[[lang=lang]]/people/+page.server.ts
@@ -1,8 +1,8 @@
 import { createMockedContributors } from '$shared/mocks';
 import type { Language } from '$shared/services/i18n';
-import { translations } from '$shared/services/i18n/translations/pages/people';
 
 import type { PageServerLoadEvent } from './$types';
+import { translations } from './_page/translation';
 
 export async function load({ params }: PageServerLoadEvent) {
   const lang = params.lang as Language;
