+++ src/routes/[[lang=lang]]/people/_page/translation/en.ts
