+++ src/routes/[[lang=lang]]/people/_page/translation/index.ts
