+++ src/routes/[[lang=lang]]/people/_page/translation/vi.ts
