+++ src/routes/[[lang=lang]]/sponsor/_page/translation/en.ts
