+++ src/routes/[[lang=lang]]/sponsor/_page/translation/index.ts
